Inline MasterContentFilter
